Update parser to parse blocks instead of doc
Parser used to parse an input as a top node type of the schema. With this change it just parses it as a block. Current parser implementation is too naive and does not actually follow the schema definition instead it was just failing to parse. Until parser is improved to handle schema definition it makes more sense to parse blocks by default.