Reimplement editing behavior to avoid doc updates
Previous implementation was updating a document content on cursor movements to insert / remove markup as needed which caused lot of needless document updates and computation.

New implementation instead revises a markdown parser so that markup nodes end up being in the document. Which are styled as invisible, but become visible when cursor is in a right position, which is accomplished by decorating edit-rang and edit-block and styling markup in those ranges as visible.

Right now block re-parse on edit is disabled but it will be enabled back again after it’s greatly simplified (as expand / collapse code staff is obsolete with this approach).